# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SentinelHealth API",
    "description": "Sistema de apoio a analises clinicas multimodais. Nao realiza diagnostico autonomo; toda classificacao e sujeita a revisao profissional.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/patients": {
      "post": {
        "tags": [
          "patients"
        ],
        "summary": "Create Patient",
        "operationId": "create_patient_patients_post",
        "parameters": [
          {
            "name": "X-Dev-Subject",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Subject"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "List Patients",
        "operationId": "list_patients_patients_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "default": true,
              "title": "Active"
            }
          },
          {
            "name": "has_analyses",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtra por pacientes com (true) ou sem (false) analise ja registrada.",
              "title": "Has Analyses"
            },
            "description": "Filtra por pacientes com (true) ou sem (false) analise ja registrada."
          },
          {
            "name": "X-Dev-Subject",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Subject"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse_PatientRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{patient_id}": {
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "Get Patient",
        "operationId": "get_patient_patients__patient_id__get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          },
          {
            "name": "X-Dev-Subject",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Subject"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "patients"
        ],
        "summary": "Update Patient",
        "operationId": "update_patient_patients__patient_id__patch",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          },
          {
            "name": "X-Dev-Subject",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Subject"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{patient_id}/observations": {
      "post": {
        "tags": [
          "patients"
        ],
        "summary": "Create Observation",
        "operationId": "create_observation_patients__patient_id__observations_post",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          },
          {
            "name": "X-Dev-Subject",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Subject"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObservationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "patients"
        ],
        "summary": "List Observations",
        "operationId": "list_observations_patients__patient_id__observations_get",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          },
          {
            "name": "X-Dev-Subject",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Subject"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ObservationRead"
                  },
                  "title": "Response List Observations Patients  Patient Id  Observations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{patient_id}/clinical-support-summary": {
      "post": {
        "tags": [
          "patients"
        ],
        "summary": "Generate Clinical Support Summary",
        "description": "Apoio a analise clinica assistido por LLM (botao \"Analisar dados\nclinicos\" abaixo do painel de alertas de anomalia). Consolida series\nde observacoes e alertas de anomalia do paciente em um resumo com\nvisao clinica, causas provaveis e direcionamento sugerido - sempre\ncomo apoio, nunca como diagnostico ou substituicao da analise do\nprofissional responsavel (ver `app.clinical_support.service`).",
        "operationId": "generate_clinical_support_summary_patients__patient_id__clinical_support_summary_post",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          },
          {
            "name": "X-Dev-Subject",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Subject"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicalSupportSummaryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/patients/{patient_id}/break-glass": {
      "post": {
        "tags": [
          "patients"
        ],
        "summary": "Create Break Glass Grant",
        "description": "Concede acesso de emergencia ao paciente por um prazo curto\n(ESCOPO_PROJETO.md secao 5.2: \"break glass\"). Nunca silencioso: gera\nevento de auditoria AUTHORIZATION imediatamente e outro em cada acesso\nrealizado sob o grant (ver `app.core.security.require_patient_access`).\n\nLimite de taxa mais restrito (`Settings.rate_limit_auth`, nao o\n`rate_limit_default` global): esta e a unica acao de elevacao de acesso\nexposta diretamente por esta API (o login/MFA em si acontece no\nCognito, fora deste backend), entao e o alvo mais provavel de abuso por\nforca bruta/enumeracao de pacientes.",
        "operationId": "create_break_glass_grant_patients__patient_id__break_glass_post",
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Patient Id"
            }
          },
          {
            "name": "X-Dev-Subject",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-Subject"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BreakGlassRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BreakGlassGrantRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PatientCreate": {
        "properties": {
          "medical_record_number": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Medical Record Number"
          },
          "full_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Full Name"
          },
          "birth_date": {
            "type": "string",
            "format": "date",
            "title": "Birth Date"
          },
          "registered_sex": {
            "type": "string",
            "maxLength": 30,
            "minLength": 1,
            "title": "Registered Sex"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 272.0,
                "minimum": 30.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          }
        },
        "type": "object",
        "required": [
          "medical_record_number",
          "full_name",
          "birth_date",
          "registered_sex"
        ],
        "title": "PatientCreate",
        "description": "Corpo de POST /patients.\n\nA idade nunca e recebida nem armazenada: e sempre derivada de\n`birth_date` (Requirement 1.3). `height_cm` e opcional (permite\ncompletar o cadastro sem a medida em maos e preencher depois via\nPATCH) - faixa fisiologicamente possivel de adulto/adolescente\n(30-272 cm, mesmo limite usado por `ObservationType.HEIGHT` em\n`app.observations.validation`, para manter as duas fontes de altura\nconsistentes)."
      },
      "PatientRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "medical_record_number": {
            "type": "string",
            "title": "Medical Record Number"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "birth_date": {
            "type": "string",
            "format": "date",
            "title": "Birth Date"
          },
          "age": {
            "type": "integer",
            "title": "Age"
          },
          "registered_sex": {
            "type": "string",
            "title": "Registered Sex"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "has_analyses": {
            "type": "boolean",
            "title": "Has Analyses",
            "default": false
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "medical_record_number",
          "full_name",
          "birth_date",
          "age",
          "registered_sex",
          "email",
          "height_cm",
          "active",
          "created_at",
          "updated_at"
        ],
        "title": "PatientRead"
      },
...
    }
  }
}